feat(lowering): define UB-safe primitive arithmetic
Haxe Int, UInt, and Float operations now lower through typed HxcIR decisions into strictly defined C. Signed wrapping, masked shifts, exceptional remainder, floating division/modulo, and Std.int boundaries no longer rely on native undefined or implementation-defined behavior.

Safe operations stay as direct, idiomatic C. Cases that need extra semantics use request-local static-inline compiler helpers, with explicit dependency closure and no hxrt feature. Unsigned intermediates are widened and narrowed per operation so C integer promotions cannot change 32-bit results.

The new arithmetic fixture compares the supported subset with Eval, compiles and runs strict C11 at O0/O2, exercises UBSan and floating divide checks, inspects optimized helper elimination, and proves runtime=none. Governance, snapshots, docs, and the primitive contract now own that evidence.

Broader statement forms, arrays, objects, strings, exceptions, general standard-library support, and runtime-backed features remain intentionally deferred to their owning beads.

+++ test/evaluation_order/expected/program.c
@@ -101,11 +101,11 @@ hxc_temp_EvaluationFixture_run_blockzx2Dlabel_conditionalzx2Djoin_n6:
   uint32_t hxc_local_EvaluationFixture_run_selected_n2 = hxc_temp_EvaluationFixture_run_conditionalzx2Dloadzx2Dresult_n9;
   hxc_field_EvaluationFixture_counter = 0;
   uint32_t hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n10 = hxc_field_EvaluationFixture_counter;
-  hxc_field_EvaluationFixture_counter = hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n10 + 1;
+  hxc_field_EvaluationFixture_counter = (uint32_t)((uint64_t)(uint32_t)hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n10 + (uint64_t)(uint32_t)1);
   uint32_t hxc_local_EvaluationFixture_run_old_n4 = hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n10;
   uint32_t hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n11 = hxc_field_EvaluationFixture_counter;
-  hxc_field_EvaluationFixture_counter = hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n11 + 1;
-  uint32_t hxc_local_EvaluationFixture_run_fresh_n5 = hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n11 + 1;
+  hxc_field_EvaluationFixture_counter = (uint32_t)((uint64_t)(uint32_t)hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n11 + (uint64_t)(uint32_t)1);
+  uint32_t hxc_local_EvaluationFixture_run_fresh_n5 = (uint32_t)((uint64_t)(uint32_t)hxc_temp_EvaluationFixture_run_incrementzx2Dloadzx2Dresult_n11 + (uint64_t)(uint32_t)1);
   uint32_t hxc_temp_EvaluationFixture_run_loadzx2Dresult_n12 = hxc_local_EvaluationFixture_run_old_n4;
   (void)hxc_temp_EvaluationFixture_run_loadzx2Dresult_n12;
   uint32_t hxc_temp_EvaluationFixture_run_loadzx2Dresult_n13 = hxc_local_EvaluationFixture_run_fresh_n5;
